PACKET-SNIFFING.MD
#Packet Sniffer
92
Network Administrators job. Network Administrators are constantly striving to maintain smooth 
operation of their networks. For this reason there is an increasing need for tools that can analyse, 
diagnose and test the functionality and the security of networks, these tools in order to perform 
their work, need to obtain the data transiting on a network, capturing it while the network is 
working. The capture process consists of obtaining, listening on the network for every transiting 
frame, independently from its source or destination. The computer sending the data is known as 
the source and the computer receiving the data is the destination. To send data over the internet, 
the source computer must first break the data into smaller pieces known as packets.
Analysis Tools
Network analysis is the process of capturing network traffic and inspecting it closely to determine 
what is happening on the network."
There are various kinds of tools dealing with the network analysis. Given the data packet and
network traffic flow information, administrators can understand network behaviour, such as application and network usage, utilization of network resources, and network anomalies.
Increasing computer networks increase the demand of network administrators which further increase the demands of packet sniffers. There are different types of passive packet sniffers used 
are: - Wireshark, TCPdump, Colasoft packet sniffer – Colasoft Capsa, etherdetect and ettercap, 
Ngrep, Snoop. In this section, key features of top 5 passive packet sniffers are discussed in detail 
i.e. Wireshark, TCPdump, Ngrep, Etherape and Snoop.